data_collection/Local-Code-Interpreter/requirements.txt
notebook==6.5.4
openai==0.27.8
gradio==3.39.0
ansi2html==1.8.0
tiktoken
Pillow